modelrisque/requirements.txt
# Kafka
kafka-python

# Database
psycopg2-binary

# ML
numpy
pandas
scikit-learn
xgboost
shap
joblib
